moved UserInvitationInfo into a different component
