+++ client/components/UserList/UserInvitationInfo/UserInvitationInfo.scss
@@ -1,0 +1,19 @@
+@import 'client/style/partials';
+
+.user-list__invitation-info {
+  position: absolute;
+  top: -25%;
+  right: 0;
+  height: 150%;
+  background-color: $ui-bg;
+  color: $text-body;
+  border: 1px solid $ui-border-dark;
+  left: -600px;
+  display: flex;
+  justify-content: space-between;
+  padding: 14px;
+
+  .icon-close {
+    cursor: pointer;
+  }
+}
