+++ client/components/UserList/UserList.scss
@@ -53,21 +53,3 @@
 
   }
 }
-
-.user-list__invitation-info {
-  position: absolute;
-  top: -25%;
-  right: 0;
-  height: 150%;
-  background-color: $ui-bg;
-  color: $text-body;
-  border: 1px solid $ui-border-dark;
-  left: -600px;
-  display: flex;
-  justify-content: space-between;
-  padding: 14px;
-
-  .icon-close {
-    cursor: pointer;
-  }
-}
