+++ client/components/UserList/UserList.tsx
@@ -6,10 +6,7 @@ import classNames from 'classnames'
 
 import { User, Users, UserStatus } from '@meta/user'
 
-import { useAppDispatch } from '@client/store'
-import { UserManagementActions } from '@client/store/userManagement'
-import { useToaster } from '@client/hooks/useToaster'
-import Icon from '@client/components/Icon'
+import UserInvitationInfo from './UserInvitationInfo'
 
 const UserColumn: React.FC<{ user: User; field: keyof User }> = ({ user, field }) => (
   <td className="user-list__cell">
@@ -23,46 +20,6 @@ const UserRoleColumn: React.FC<{ user: User }> = ({ user }) => {
     <td className="user-list__cell">
       <div className="user-list__cell--read-only">{i18n.t<string>(Users.getI18nRoleLabelKey(user.roles[0].role))}</div>
     </td>
-  )
-}
-
-const UserInvitationInfo: React.FC<{ user: User; onClose: () => void }> = ({ user, onClose }) => {
-  const dispatch = useAppDispatch()
-  const { i18n } = useTranslation()
-  const { toaster } = useToaster()
-
-  return (
-    <div className="user-list__invitation-info">
-      <div>
-        <div>
-          {`${i18n.t('userManagement.invitationLink')}: ${window.location.origin}/login?invitationUuid=${
-            user.roles[0].invitationUuid
-          }`}
-        </div>
-        <div>
-          <button
-            className="btn-s btn-link"
-            onClick={async () => {
-              dispatch(
-                UserManagementActions.sendInvitationEmail({
-                  countryIso,
-                  invitationUuid: user.roles[0].invitationUuid,
-                })
-              ).then(() => {
-                toaster.success(i18n.t<string>('userManagement.invitationEmailSent'))
-                onClose()
-              })
-            }}
-            type="button"
-          >
-            {i18n.t<string>('userManagement.sendInvitation')}
-          </button>
-        </div>
-      </div>
-      <div onClick={onClose} role="button" tabIndex={0} aria-hidden="true">
-        <Icon name="remove" />
-      </div>
-    </div>
   )
 }
